move querylog consumer out of arroyo
