+++ rust_snuba/rust_arroyo/src/processing/mod.rs
@@ -1,5 +1,4 @@
 pub mod strategies;
-pub mod querylog_processor;
 
 use crate::backends::{AssignmentCallbacks, Consumer};
 use crate::types::{Message, Partition, Topic};
