+++ rust_snuba/src/consumers/mod.rs
