+++ rust_snuba/src/consumers/querylog_consumer.rs
@@ -11,11 +11,12 @@ use rust_arroyo::backends::Consumer as ConsumerTrait;
 use rust_arroyo::backends::kafka::CustomContext;
 use rust_arroyo::backends::kafka::KafkaConsumer;
 use rust_arroyo::backends::kafka::config::KafkaConfig;
-use rust_arroyo::processing::querylog_processor;
-use rust_arroyo::processing::querylog_processor::RawQueryLogKafkaJson;
 use rust_arroyo::types::Partition;
 use rust_arroyo::types::Topic;
 use rust_arroyo::utils::clickhouse_client::ClickhouseClient;
+
+use crate::processors::querylog_processor::RawQueryLogKafkaJson;
+use crate::processors::querylog_processor;
 
 
 const TIMEOUT_MS: Duration = Duration::from_millis(2000);
