+++ rust_snuba/src/lib.rs
@@ -1,1 +1,2 @@
-pub mod consumer;
+pub mod consumers;
+pub mod processors;
