+++ rust_snuba/src/main.rs
@@ -1,5 +1,6 @@
 extern crate rust_arroyo;
 pub mod storage;
+pub mod processors;
 
 fn main() {
     println!("Hello, world!");
